Feat: [#5] GeoUtil 만들기
# Conflicts:
#	src/main/java/umc/wegg/service/PlanService/PlanQueryServiceImpl.java
