+++ src/main/java/umc/wegg/service/PlanService/PlanQueryServiceImpl.java
@@ -7,6 +7,7 @@ import umc.wegg.domain.User;
 import umc.wegg.dto.PlanResponseDTO;
 import umc.wegg.repository.PlanRepository;
 import umc.wegg.repository.UserRepository;
+import umc.wegg.util.GeoUtil;
 
 import java.util.List;
 
@@ -45,7 +46,7 @@ public class PlanQueryServiceImpl implements PlanQueryService {
         double userLon = user.getCurrentLon();
 
         // 경계 확인 (500미터 이내인지 확인하는 메서드 호출)
-        boolean isWithinBoundary = GeoUtils.isWithinPlanBoundary(planLat, planLon, userLat, userLon, 0.5);  // 0.5km (500m)
+        boolean isWithinBoundary = GeoUtil.isWithinPlanBoundary(planLat, planLon, userLat, userLon, 0.5);  // 0.5km (500m)
 
         // 결과에 맞는 메시지를 설정
         String message = isWithinBoundary ? "장소 인증에 성공했습니다!" : "장소 인증에 실패했습니다!";
@@ -53,23 +54,7 @@ public class PlanQueryServiceImpl implements PlanQueryService {
         // LocationVerificationResponseDTO 반환
         return new PlanResponseDTO.LocationVerificationResponseDTO(message);
     }
-    public class GeoUtils {
 
-        private static final double EARTH_RADIUS = 6371; // 지구 반지름 (킬로미터)
-
-        // 사용자가 계획 범위 내에 있는지 판단하는 메서드
-        public static boolean isWithinPlanBoundary(double planLat, double planLon, double userLat, double userLon, double radius) {
-            // 위도와 경도 차이로 거리 계산 (킬로미터 단위)
-            double latDistance = Math.toRadians(userLat - planLat) * EARTH_RADIUS;
-            double lonDistance = Math.toRadians(userLon - planLon) * EARTH_RADIUS * Math.cos(Math.toRadians(planLat));
-
-            // 두 위치 사이의 거리 계산
-            double distance = Math.sqrt(latDistance * latDistance + lonDistance * lonDistance);
-
-            // 사용자가 범위 내에 있으면 true
-            return distance <= radius;
-        }
-    }
 }
